declutter App.js
